-added expiration date field in stock form -added low stock amount field in SKU MODEL

+++ Shopper/Services/Implementations/ProductService.cs
@@ -128,6 +128,8 @@ namespace Shopper.Services.Implementations
                 .Include(p => p.Skus)
                 .ThenInclude(sku => sku.SkuAttributes)
                 .ThenInclude(skuAtt => skuAtt.Option)
+                .Include(p => p.Skus)
+                .ThenInclude(s => s.Expiration)
                 .FirstAsync(p => p.Id.Equals(productId));
             product.Skus = product.Skus.OrderByDescending(sku => sku.CreatedAt).ToList();
             return product;
@@ -187,8 +189,9 @@ namespace Shopper.Services.Implementations
             return prod.Entity;
         }
 
-        public async Task<Sku> AddProductToStockAsync(Sku sku, List<ushort> attributeOptions)
+        public async Task<Sku> AddProductToStockAsync(SkuViewFormModel formModel, List<ushort> attributeOptions)
         {
+            var sku = formModel.Sku;
             sku.RemainingQuantity = sku.Quantity;
             sku.IsOnSale = false;
             var newSku = _dbContext.Skus.Add(sku);
@@ -201,14 +204,24 @@ namespace Shopper.Services.Implementations
                 await _dbContext.SkuAttributes.AddRangeAsync(skuAttributes);
             }
 
+            if (formModel.ExpirationDate != null)
+            {
+                sku.Expiration = new Expiration
+                {
+                    SkuId = sku.Id,
+                    ExpirationDate = formModel.ExpirationDate.Value,
+                };
+            }
+
             await _dbContext.SaveChangesAsync();
             return newSku.Entity;
         }
 
-        public async Task<Sku> UpdateStockItemAsync(Sku sku,Sku updated, List<ushort> attributeOptionIds)
+        public async Task<Sku> UpdateStockItemAsync(Sku sku,SkuViewFormModel formModel, List<ushort> attributeOptionIds)
         {
             try
             {
+                var updated = formModel.Sku;
                 await _dbContext.Database.BeginTransactionAsync();
                 sku.Quantity = updated.Quantity;
                 sku.SellingPrice = updated.SellingPrice;
@@ -216,11 +229,29 @@ namespace Shopper.Services.Implementations
                 sku.MaximumDiscount = updated.MaximumDiscount;
                 sku.Date = updated.Date;
                 await _dbContext.Entry(sku).Collection(s => s.SkuAttributes).LoadAsync();
-                _dbContext.SkuAttributes.RemoveRange(sku.SkuAttributes);
-                var skuAttributes = attributeOptionIds
-                    .Select(attributeOption => new SkuAttribute
-                        {Sku = sku, AttributeOptionId = attributeOption}).ToList();
-                await _dbContext.SkuAttributes.AddRangeAsync(skuAttributes);
+
+                if (sku.SkuAttributes.Any())
+                {
+                    _dbContext.SkuAttributes.RemoveRange(sku.SkuAttributes);
+                }
+
+                if (!attributeOptionIds.IsNullOrEmpty())
+                {
+                    var skuAttributes = attributeOptionIds
+                        .Select(attributeOption => new SkuAttribute
+                            {Sku = sku, AttributeOptionId = attributeOption}).ToList();
+                    await _dbContext.SkuAttributes.AddRangeAsync(skuAttributes);
+                }
+
+                if (formModel.ExpirationDate != null)
+                {
+                    await _dbContext.Expirations.AddAsync(new Expiration
+                    {
+                        SkuId = sku.Id,
+                        ExpirationDate = formModel.ExpirationDate.Value,
+                    });
+                }
+
                 var updatedSku = _dbContext.Skus.Update(sku);
                 await _dbContext.SaveChangesAsync();
                 _dbContext.Database.CommitTransaction();
